# Conflicts:
#	AssetBundles/Local/iOS/dragon_dark_local
#	AssetBundles/Local/iOS/npc_medieval_village_castle_dark
#	AssetBundles/Local/iOS/shared
#	AssetBundles/Remote/iOS/dragon_dark
#	AssetBundles/iOS/dragon_dark.manifest
#	AssetBundles/iOS/dragon_dark_local.manifest
#	AssetBundles/iOS/iOS
#	AssetBundles/iOS/iOS.manifest
#	AssetBundles/iOS/npc_medieval_village_castle_dark.manifest
#	AssetBundles/iOS/shared.manifest
#	Assets/Editor/Addressables/generated/Android/addressablesCatalog.json
#	Assets/Editor/Addressables/generated/editor_addressablesCatalog.json
#	Assets/Editor/Addressables/generated/iOS/addressablesCatalog.json
#	Assets/Editor/Addressables/generated/iOS/assetBundlesCatalog.json
#	Assets/Editor/Addressables/generated/iOS/downloadablesCatalog.json
#	Assets/Resources/Rules/entityDefinitions.xml
#	Docs/Content/HungryDragonContent_Gameplay.xlsx
#	Docs/Content/xml/entityDefinitions.xml


Former-commit-id: 5025d8d3db2382e51d3e76afd4aa249226cf2d64

+++ Assets/Game/Gameplay/Entities/IEntity.cs
@@ -60,7 +60,7 @@ abstract public class IEntity :  MonoBehaviour, ISpawnable {
 	protected float m_health;
 	public float health { get { return m_health; } set { m_health = value; } }
 
-	public virtual int score { get { return 0; } }
+	public virtual float score { get { return 0; } }
 
 	public ISpawnable[] m_otherSpawnables;
 	protected int m_otherSpawnablesCount;
